requirements.txt
fastapi==0.83.0
uvicorn[standard]==0.18.3
starlette==0.19.1
Jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
pymongo[srv]==4.2.0
python-dotenv
-e .